Minor action, actionslists adjustments
