--- Assets/_Scripts/Actions/ActionCost.cs
@@ -1,11 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-[System.Serializable]
-public class ActionCost {
-    public int HP;
-    public int mana;
-    public int PP;
-    public int stamina;
-}
