--- Assets/_Scripts/Actions/ActionCost.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2d6aa186485111848ad6455785443f91
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
